Create Delegate Call
This contract needs to be deployed and the getPWNsig needs calling. This will bring up an address. Coy this address and then in the Ethernaut console type
await getTransaction({from:player, to:instance, data:"enter add address here")}
This should change you to the owner and you can complete the level.