Update King
This .sol needs deploying from the injected 3 web. You need to add the contract address into the address field. Make a 1 ETH payment also. The amount of the first place can be seen in the console by typing: await contract.prize().  1 ETH is sent ,as this is on the test network and is higher than the current 1st place amount. To check the address has changed, you can type: await.contract._king()     This should give you your address.