Create Re-entrancy
You need these two contracts, as the attack contract calls upon the Reentrance.sol. Compile both. If you check the      await getBalance(contract.address)      in the console, you are given a figure of 0.001. This equates to 1 Finney. because we can only send this amount, we need to operate in Finney. If you deploy the attack contract and send 1 Finney then the balances are equal which means we can then transact the donate to target and steal all the funds.